+++ packages/dictionary/src/ReadonlyDictionary.ts
@@ -1,8 +1,8 @@
 import { BinaryPredicate, Cloneable, Mapping } from '@jamashita/anden-type';
 import { Collection } from '@jamashita/lluvia-collection';
 
-export interface ReadonlyProject<out K, out V> extends Collection<K, V>, Cloneable<ReadonlyProject<K, V>> {
-  filter(predicate: BinaryPredicate<V, K>): ReadonlyProject<K, V>;
+export interface ReadonlyDictionary<out K, out V> extends Collection<K, V>, Cloneable<ReadonlyDictionary<K, V>> {
+  filter(predicate: BinaryPredicate<V, K>): ReadonlyDictionary<K, V>;
 
   has(key: K): boolean;
 
@@ -10,7 +10,7 @@ export interface ReadonlyProject<out K, out V> extends Collection<K, V>, Cloneab
 
   keys(): Iterable<K>;
 
-  map<W>(mapping: Mapping<V, W>): ReadonlyProject<K, W>;
+  map<W>(mapping: Mapping<V, W>): ReadonlyDictionary<K, W>;
 
   toMap(): Map<K, V>;
 }