+++ packages/dictionary/src/mock/MockDictionary.ts
@@ -1,35 +1,35 @@
 import { UnimplementedError } from '@jamashita/anden-error';
 import { Quantity } from '@jamashita/lluvia-collection';
-import { AProject } from '../AProject';
+import { ADictionary } from '../ADictionary';
 
-export class MockProject<out K, out V> extends AProject<K, V> {
-  public constructor(project: Map<K, V>) {
+export class MockDictionary<out K, out V> extends ADictionary<K, V> {
+  public constructor(dictionary: Map<K, V>) {
     const map: Map<K | string, [K, V]> = new Map();
 
-    project.forEach((v: V, k: K) => {
+    dictionary.forEach((v: V, k: K) => {
       map.set(Quantity.genKey(k), [k, v]);
     });
 
     super(map);
   }
 
-  public duplicate(): MockProject<K, V> {
+  public duplicate(): MockDictionary<K, V> {
     throw new UnimplementedError();
   }
 
-  public filter(): MockProject<K, V> {
+  public filter(): MockDictionary<K, V> {
     throw new UnimplementedError();
   }
 
-  public map<W>(): MockProject<K, W> {
+  public map<W>(): MockDictionary<K, W> {
     throw new UnimplementedError();
   }
 
-  public remove(): MockProject<K, V> {
+  public remove(): MockDictionary<K, V> {
     throw new UnimplementedError();
   }
 
-  public set(): MockProject<K, V> {
+  public set(): MockDictionary<K, V> {
     throw new UnimplementedError();
   }
 }